codec: json: try fast-path for decoding floats before defaulting to strconv.ParseFloat
The json format for floats is more constrained that what strconv supports.
We noticed some slowdown with float parsing which we have attempted to resolve
by walking through the text and doing an exact parse if within the narrow constraints where
it is simple.

Also, have float32 not convert back and forth between float64 unnecessarily.

+++ codec/decode.go
@@ -2900,11 +2900,11 @@ func (d *Decoder) NumBytesRead() int {
 // else if will call DecodeFloat64 and ensure the value doesn't overflow.
 //
 // Note that we return float64 to reduce unnecessary conversions
-func (d *Decoder) decodeFloat32() float64 {
+func (d *Decoder) decodeFloat32() float32 {
 	if d.js {
 		return d.jdec.DecodeFloat32() // custom implementation for 32-bit
 	}
-	return chkOvf.Float32V(d.d.DecodeFloat64())
+	return float32(chkOvf.Float32V(d.d.DecodeFloat64()))
 }
 
 // ---- container tracking
